Add KeywordFacetLabel component
